큐엔에이 합침
