--- delivery_eyo/src/main/java/com/jhta/delivery/controller/QnaController.java
@@ -1,11 +1,0 @@
-package com.jhta.delivery.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-@Controller
-public class QnaController {
-	@RequestMapping("/member/qna")
-	public String qna() {
-		return ".member.qna";
-	}
-}
