+++ delivery_eyo/src/main/java/com/jhta/delivery/controller/QnaListController.java
@@ -7,6 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
 import com.jhta.delivery.service.QnaService;
 import com.jhta.delivery.vo.QnaVo;
@@ -20,9 +22,6 @@ public class QnaListController {
 	public String qnaList(Model model,@RequestParam(value = "pageNum",defaultValue = "1")int pageNum,
 		String field,String keyword,int memberNum) {
 		HashMap<String , Object> map = new HashMap<String, Object>();
-		//검색조건 map에 담기
-		map.put("field", field);
-		map.put("keyword", keyword);
 		
 		//페이지 유틸 사용하기
 		int totalRowCount=service.count(map);
@@ -31,16 +30,46 @@ public class QnaListController {
 		map.put("endRow", pu.getEndRow());
 		map.put("memberNum", memberNum);
 		
-		System.out.println(memberNum);
+	
 		List<QnaVo> list= service.qnaList(map);
 		model.addAttribute("startRow",pu.getStartRow());
 		model.addAttribute("endRow",pu.getEndRow());
 		model.addAttribute("list",list);
 		model.addAttribute("pu",pu);
-		model.addAttribute("field",field);
-		model.addAttribute("keyword",keyword);
 		model.addAttribute("memberNum",memberNum);
-		
+
 		return ".member.QnaList";
+	}
+	
+	@GetMapping("/member/qnaDetail")
+	public String qnadetail(int num,Model model) {
+		QnaVo vo=service.detail(num);//상세글 정보
+		QnaVo next = service.next(num);//다음글
+		QnaVo pre = service.pre(num);//이전글
+		model.addAttribute("vo",vo);
+		model.addAttribute("next",next);
+		model.addAttribute("pre",pre);
+		return ".member.qnaDetail";
+		
+	}
+	@RequestMapping("/member/qna")
+	public String qna() {
+		return ".member.qna";
+	}
+	@RequestMapping(value = "/member/personalQna",method=RequestMethod.GET)
+	public String qnainsert() {
+		return ".member.personalQna";
+	}
+	@RequestMapping(value = "/member/personalQna",method=RequestMethod.POST)
+	public String qnainsertOk(QnaVo vo) {
+		try {
+			service.insert(vo);
+			//성공하면 다시 리스트로 돌아가도록
+			return ".member.QnaList";
+		} catch (Exception e) {
+			// TODO: handle exception
+			System.out.println(e.getMessage());
+			return "error";
+		}
 	}
 }
