+++ delivery_eyo/src/main/java/com/jhta/delivery/dao/Qna_ImgDao.java
@@ -1,0 +1,13 @@
+package com.jhta.delivery.dao;
+
+import org.mybatis.spring.SqlSessionTemplate;
+import org.springframework.stereotype.Repository;
+import com.jhta.delivery.vo.Qna_ImgVo;
+@Repository
+public class Qna_ImgDao {
+	private SqlSessionTemplate sqlSession;
+	private final String NAMESPACE="mybatis.QnaImgMapper";
+	public int insert(Qna_ImgVo vo) {
+		return sqlSession.insert(NAMESPACE+".insert",vo);
+	}
+}
