+++ delivery_eyo/src/main/java/com/jhta/delivery/service/QnaService.java
@@ -4,12 +4,18 @@ import java.util.HashMap;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
 import com.jhta.delivery.dao.QnaDao;
+import com.jhta.delivery.dao.Qna_ImgDao;
 import com.jhta.delivery.vo.QnaVo;
+import com.jhta.delivery.vo.Qna_ImgVo;
 @Service
 public class QnaService {
 	@Autowired
 	private QnaDao dao;
+	@Autowired
+	private Qna_ImgDao dao1;
 	public int count(HashMap<String, Object>map) {
 		return dao.count(map);
 		
@@ -32,6 +38,11 @@ public class QnaService {
 	}
 	public int insert(QnaVo vo) {
 		return dao.insert(vo);
+		
+	}
+	@Transactional
+	public int insert(Qna_ImgVo vo) {
+		return dao1.insert(vo);
 		
 	}
