+++ delivery_eyo/src/main/java/com/jhta/delivery/vo/Qna_ImgVo.java
@@ -1,0 +1,19 @@
+package com.jhta.delivery.vo;
+
+public class Qna_ImgVo {
+	// 이미지번호 
+    private int num;
+
+    // 이미지경로 
+    private String imgPath;
+
+    // 원본파일명 
+    private String originalFilename;
+
+    // 세이브파일명 
+    private String saveFilename;
+
+    // 문의글번호 
+    private int inquiryNum;
+
+}
