+++ delivery_eyo/src/main/resources/com/jhta/delivery/mybatis/mapper/QnaImgMapper.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE mapper 
+PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+  "http://www.ibatis.com/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="mybatis.QnaImgMapper">
+	<insert id="insert" parameterType="QnaImg">
+		insert into INQUIRY_IMG values(inquiry_img_seq.nextval,#{imgPath},#{originalFilename},#{inquiryNum})
+	</insert>
+</mapper>
